convert sensor globals and serial packet to integer types
Replace float/uint16 sensor globals (tps, fps_bar, bat_v) and the
TPS axis with compact integer types to eliminate floating-point
arithmetic from all sensor read paths.

- readTPS: uint16 per-mille → uint8 percent (0–100)
- readFPS: float bar → uint8 1/8-bar units (0–80)
- readBatV: float V → uint8 1/16-V units (0–255)
- tps_axis: uint16 per-mille → uint8 percent; EEPROM magic bumped
  to 0xAC to force re-init; layout shrinks from 200 to 195 bytes
- Sensor packet shrinks from 16 to 13 bytes accordingly
- protocol.py decode updated to match new struct layout and scales

+++ asf_efi.h
@@ -5,11 +5,11 @@
 #include "sensors.h"
 #include "injection.h"
 
-extern uint16_t tps;       // 0–1000 per-mille (0 = 0%, 1000 = 100%)
-extern float    fps_bar;
+extern uint8_t  tps;       // 0–100 percent
+extern uint8_t  fps_bar;   // 0–80, units = 0.125 bar
 extern int16_t  iat_degc;  // whole °C
 extern int16_t  et_degc;   // whole °C
-extern float    bat_v;     // battery voltage in V
+extern uint8_t  bat_v;     // battery voltage in 1/16 V units (0.0625 V per count)
 
 extern volatile uint16_t rpm;
 extern volatile bool     pump_active;
@@ -18,7 +18,7 @@ extern bool              pump_mode_always_on; // false = PID (default), true = f
 extern uint8_t           pump_pwm;           // last analogWrite value to pump (0–255)
 
 extern uint16_t rpm_axis[RPM_BINS];
-extern uint16_t tps_axis[TPS_BINS];   // 0–1000 per-mille breakpoints
+extern uint8_t  tps_axis[TPS_BINS];   // 0–100 percent breakpoints
 
 extern uint16_t inj_map[RPM_BINS][TPS_BINS];
 extern uint16_t iat_correction[IAT_CORR_BINS];  // Q8.8: 256 = 1.0 (no correction)
